[docs] trying to resolve failing tests
